requirements.txt
# ============================================================
#  requirements.txt
#  설치 명령: pip install -r requirements.txt
# ============================================================

flask==3.0.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0

# MySQL 사용 시 (PythonAnywhere 운영 배포):
# pymysql==1.1.0
